+++ code/src/sixsq/nuvla/server/resources/configuration_nuvla.clj
@@ -7,14 +7,15 @@ default values.
   (:require
     [clojure.tools.logging :as log]
     [environ.core :as env]
+    [sixsq.nuvla.pricing.impl :as pricing-impl]
     [sixsq.nuvla.server.resources.common.crud :as crud]
     [sixsq.nuvla.server.resources.common.std-crud :as std-crud]
     [sixsq.nuvla.server.resources.common.utils :as u]
     [sixsq.nuvla.server.resources.configuration :as p]
     [sixsq.nuvla.server.resources.configuration-template :as ct]
     [sixsq.nuvla.server.resources.configuration-template-nuvla :as tpl-nuvla]
-    [sixsq.nuvla.server.resources.pricing.stripe :as stripe]
     [sixsq.nuvla.server.resources.spec.configuration-template-nuvla :as ct-nuvla]
+    [sixsq.nuvla.server.util.namespace-utils :as dyn]
     [sixsq.nuvla.server.util.response :as r]))
 
 
@@ -58,13 +59,18 @@ default values.
                                         crud/retrieve-by-id-as-admin
                                         :stripe-api-key)
                                     (env/env :stripe-api-key))]
-        (stripe/set-api-key! stripe-api-key)
-        (alter-var-root #'*stripe-api-key* (constantly stripe-api-key)))
-      (when-let [stripe-client-id (or (-> config-instance-url
-                                          crud/retrieve-by-id-as-admin
-                                          :stripe-client-id)
-                                      (env/env :stripe-client-id))]
-        (alter-var-root #'*stripe-client-id* (constantly stripe-client-id)))
+        (let [pricing-instance (some-> "sixsq.nuvla.pricing.stripe.stripe" dyn/load-ns pricing-impl/set-impl!)]
+          (if (some? pricing-instance)
+            (do
+              (pricing-impl/set-impl! pricing-instance)
+              (pricing-impl/set-api-key! stripe-api-key)
+              (alter-var-root #'*stripe-api-key* (constantly stripe-api-key))
+              (when-let [stripe-client-id (or (-> config-instance-url
+                                                  crud/retrieve-by-id-as-admin
+                                                  :stripe-client-id)
+                                              (env/env :stripe-client-id))]
+                (alter-var-root #'*stripe-client-id* (constantly stripe-client-id))))
+            (log/error "Stripe-api-key configured but no princing implementation found!"))))
       (catch Exception e
         (log/error (str "Exception when loading Stripe api-key/client-id: " e))))))