+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+  - Nuvla api test jar artifact
+
 ### Changed
+
+  - Dynamic load namespace helper
+  - Pricing related resources moved to another artifact
 
 ## [5.18.0] - 2021-10-28